Updated styles
Mostly changed the styles of many different elements in the project. Added/changed a few details to console.log comments in the .js file. Added some classes to elements to better target them with css. Overall, project is complete but not fully functioning. See .docx file added soon for details.